requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
whitenoise==5.2.0
dj_database_url==0.5.0
blessings==1.7
boto==2.49.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.0
colorama==0.4.4
colorlog==5.0.1
cryptography==3.4.8
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
django-crispy-forms==1.11.2
django-markwhat==1.6.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
dnspython==2.1.0
h5py==3.2.1
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
pymongo==3.12.0
PyRect==0.1.4
PyTweening==1.0.3
pytz==2021.1
requests-toolbelt==0.7.0
soupsieve==2.2.1
sqlparse==0.2.4
termcolor==1.1.0
django-heroku==0.3.1
gunicorn==20.0.4